fix(stats/controller): rttInfo narrow, GC-pressure removal, combo cancel race
C4 (MediaCodecDecoderRenderer): networkLatencyMs cast '(float)(rttInfo
    >> 32)' kept the long sign bit and produced wrong magnitude when
    the upper word's high bit was set. The existing perf-overlay path
    at lines 1811/1869 narrows to (int) first to match the wire pack
    format. New stats path now matches: '(int)(rttInfo >> 32)'.

H7+H8 (StatsOverlay): the 'lightweight, single-allocation StringBuilder'
    comment was misleading — every redraw built 8 strings via 8
    .toString() calls plus String.format('%.1f ms', v) inside each
    line, which allocated a Formatter per call. Eight format-allocs
    per redraw at 60fps = ~480 throwaway Formatter+intermediate-string
    objects per second on every single frame.

    Fix: introduce appendOneDecimal(StringBuilder, float, String)
    helper with half-up rounding and NaN/Infinity guards. Replace
    every String.format with manual decimal formatting via append
    chains. Add per-stat dirty caches (NaN-safe via
    Float.floatToRawIntBits comparison); updateClientStats /
    updateServerStats / updateWifiStats early-return when nothing
    changed and skip invalidate() entirely when mode == OFF.
    toggle()/setMode() force a repaint via clearDirtyCaches().

    Comment rewritten to be honest: StringBuilder reused across
    redraws to avoid GC pressure from repeated 'new StringBuilder()',
    not to avoid all allocation — toString() still allocates a String
    per distinct text label. The win is in avoiding the auto-grow
    path of fresh builders, not in zero allocation.

H1+H2 (ControllerHandler): two related bugs in the Select+L1 long-press
    that toggles the stats overlay.

    H1 race: cancel logic at the original two sites evaluated BEFORE
    inputMap was fully resolved this frame, but the post-runnable
    was scheduled by an earlier frame. A user releasing Select or L1
    mid-frame could have the post-runnable fire even though the
    cancel check saw the post-release state.

    H2 collision: BACK + LB is a standard 'previous weapon' / 'swap
    weapon' combo in many shooters. Holding 2s for the stats toggle
    also flowed those button events to the host = unintended in-game
    effect.

    Fix: extract the cancel check into a single helper
    cancelStatsOverlayHoldIfComboBroken(currentInputMap) that runs
    AFTER inputMap is fully resolved for the frame; called from both
    handleButtonUp and handleButtonDown's combo-breaking path.
    Existing dispose-time mainThreadHandler.removeCallbacks retained.

    H2 surface: added DEFAULT_STATS_OVERLAY_COMBO constant +
    statsOverlayCombo field with TODO referencing a future
    PreferenceConfiguration entry (no settings UI added in this
    commit; user can change combo in settings once that lands).

+++ app/src/main/java/com/limelight/binding/video/MediaCodecDecoderRenderer.java
@@ -1883,7 +1883,10 @@ public class MediaCodecDecoderRenderer extends VideoDecoderRenderer implements C
                 // Also send structured stats for the SudoVDA stats overlay
                 float totalTimeMs = lastTwo.totalFrames > 0 ?
                         (float)lastTwo.totalTimeMs / lastTwo.totalFrames : 0;
-                float networkLatencyMs = (float)(rttInfo >> 32); // average RTT
+                // Narrow to int first (matching the perf-overlay path above) before widening to float.
+                // The pack format stores avg RTT in the high 32 bits; casting the long directly to
+                // float can produce a wrong magnitude when the high bit of the upper word is set.
+                float networkLatencyMs = (int)(rttInfo >> 32); // average RTT
                 String codecName;
                 if ((videoFormat & MoonBridge.VIDEO_FORMAT_MASK_H264) != 0) {
                     codecName = "H.264";
